nfs4: client fix int vs long when constructing layoutget requests
Target: master
Acked-by: Karsten Schwank

+++ core/src/main/java/org/dcache/nfs/v4/client/CompoundBuilder.java
@@ -538,7 +538,7 @@ public class CompoundBuilder {
     }
 
     public CompoundBuilder withLayoutget(boolean signal_layout_avail,
-            int layout_type, int iomode, int offset, int length, int minlength,
+            int layout_type, int iomode, long offset, long length, long minlength,
             int maxcount, stateid4 stateid) {
         nfs_argop4 op = new nfs_argop4();
         op.argop = nfs_opnum4.OP_LAYOUTGET;
